Added geocodes to Chapter 1
I added geocodes to every place I could find in Chapter 1. Most of the locations are actually visited at some point by Stephen or other characters. Many of the locations that had footnotes in the book, and I referenced them for accuracy. 

However, I was unsure whether to include some locations that are indirectly mentioned, but I did for good measure. These locations are Tullamore on p.36, a type of tobacco that is imported from Tullamore, Ireland.  Eton on p. 29, since Stephen wears an Eton jacket, the style originally from Eton, England, and p. 35 The Paris Funds. I also included more farflung places such as the Mozambique Channel, the Mississippi River (indirectly mentioned), Prague, Greece, and Rome since they were mentioned in Stephen's thoughts or conversations. 

The one location I could not find online was the Alleghanies on p. 35. Dante supposedly went to a convent there. It seems to be a mispelled version of the Alleghenies, and the Alleghenies are in North America, which I doubt is the actual location, so I did not include it. Therefore, I did not geocode this.