Add names to Chapter 1 (until l.1000), for issue #37
I decided to mark up uncle Charles as Uncle Charles, instead of just Charles, as Stephen kept referring to him as Uncle Charles, as if Uncle was his title, like Father Arnell or Brother Michael, which I thought was quite interesting. 
I also decided to mark up the world and the universe as place names in the passage where Stephen situates himself in the universe, going from Conglowes to "The Universe", as it seemed significant as a place to the passage. 
Additionally, I did not only mark up the names of characters, but also other people, e.g. Parnell, as I thought they were still persons of equal importance to the novel.