Add names to Chapter 1 - paragraph about God
I decided to mark up "God" as a personal name in the paragraph where Stephen thinks about how "God" is the name of God, even thought it is not a personal name in general. In that paragraph, however, God becomes a person with a name, which could be interesting to analyse for anyone using the mark ups of names for their analysis of Portrait.