Mark up names Chapter 1 l.1000-1119 #37
I noticed, as the Christmas dinner heated, there was a sharp uptake in the frequency of names due to the characters' yelling at each other. The recurring line "Simon!" especially struck me as the association with simony kept popping in my head (thanks to it being one of the three words in "The Dead," and often mentioned by Professor Cole.