Add dialog attribution to Chapter 3, for issue #7
I had a few questions/judgment decisions: 
I wondered how to attribute dialog to unnamed characters (EX: the woman who gives Stephen directions to church near the end of the chapter). For some of them, I gave an attribution (EX: for the confession at the end of the chapter, I labeled the priest Stephen is confessing to, "Priest"). 

Also, in the first dialog of the chapter I found it hard to identify who exactly is speaking (so I did not attribute these lines).

I also assumed that the rector is still "Father Conmee" (even though this full name is not referenced at all in Chapter 3). 

Before the sermons, there was a prayer said. It is not explicitly stated that the pastor ("Father Arnall") gives this opening prayer, but there is no indication of anyone else.

Some of the imaginary dialog in Stephen's head I was not sure how to attribute, but for some I did attribute (EX: I labeled "Virgin Mary" as speaking in the imaginary scene with Stephen and Emma)

Finally, I gather that the latin words spoken at end of chapter of the communion are spoken by the priest. First issue: which priest? (Or what name? Retreat is over; is it still Arnall?) Then, after "Amen" there is one more line of dialogue repeating "Corpus Domini nostri", was not sure who to attribute this to either; so I did not attribute these final three lines of dialog.