More Ch. 5 Geocodes for Issue #43
The nun's madhouse is St. Vincent's Psychiatric Hospital, Fairview. For the city on page 190, I just put Stephen's new address that was in the footnotes, which is 8 Royal Terrace in Fairview, because I think it would be important to put somewhere even though his new home is not explicitly mentioned. Hopkins' corner on p.191 is supposed to be somewhere on O'Connell Street. Not sure if the Adelphi hotel I found online is the same Adelphi hotel. On p. 267, I could not find a place called Larras in Ireland, but I found an obscure place possibly called Iarras-Domhnan. Would that be a mispelling? I will finish geocoding the rest of Chapter 5 in the future, so I will just submit a pull request now. Ch. 3 and Ch. 4 contain fewer location names.