Finished geocoding Ch. 4. Issue #43
I finished geocoding for the entire novel besides Ch. 2, which someone else coded for. I excluded very far away and general place names such as "Europe", and some repetitions of places that I already coded for. However, I did code Norfolk, England for the Norfolk coat, since it is interesting to see how British styles of dress are noted on the Irish (there is an Eton jacket mentioned in another instance). Is it due to possible colonialism? Additionally, I found out that Ballybough Bridge (in the footnotes), the bridge over the Tolka river, is now named Luke Kelly Bridge, which I coded for.